Move KonashiEvent to events package
