+++ Konashi/konashi-v2-android-sdk/src/main/java/com/uxxu/konashi/lib/events/KonashiDigitalEvent.java
@@ -1,6 +1,5 @@
 package com.uxxu.konashi.lib.events;
 
-import com.uxxu.konashi.lib.KonashiEvent;
 import com.uxxu.konashi.lib.listeners.KonashiBaseListener;
 import com.uxxu.konashi.lib.listeners.KonashiDigitalListener;
